+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 <!-- markdownlint-disable MD041 MD036 -->
-## 4.9.6
+## 4.10.0
 
 ### 📦函数